Check XM tag (number of mismatches) in case of same CIGAR string or CIGAR string with same number of matches.
Check XM tag (number of mismatches) in case of same CIGAR string
or CIGAR string with same number of matches to determine to which
haplotype a read maps the best. Bowtie2 can report the same CIGAR
string even one without mismatches, but still indicating a mismatch
in the XM tag.